refactor:#67 감자 svg 변경
